Added PopH264_Version to all builds (was stopping windows from loading)

+++ Source/PopH264_Version.cpp
@@ -50,9 +50,9 @@ namespace PopH264
 
 	//	gr: use macros once linux is automated
 	//const Soy::TVersion	Version(VERSION_MAJOR,VERSION_MINOR,VERSION_PATCH);
-	constexpr size_t	VersionMajor = 1;
-	constexpr size_t	VersionMinor = 6;
-	constexpr size_t	VersionPatch = 0;
+	constexpr int	VersionMajor = 1;
+	constexpr int	VersionMinor = 6;
+	constexpr int	VersionPatch = 0;
 }
